re orginize the code
