+++ index.js
@@ -1,0 +1,3 @@
+export {isShabbat} from './package/isShabbat.mjs';
+export {useShabbatCheck} from './package/shabbatCheck.mjs';
+export {shabbatDate} from './package/shabbatDate.mjs';
