+++ static/shabbatPage.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Shabbat Kodesh</title>
+</head>
+<body>
+<h1 style="text-align: center;"> !!! שבת קודש</h1>
+<img style="width: 100%;"
+     src="https://www.kipa.co.il/userFiles/735-415/296656_373af8629a50ff540978724a33c9d96a.jpg" alt="shabbat kodesh">
+</body>
+
+</html>
